requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
requests>=2.31.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.1
